Created multi user logic
Now we can create a company or a client when creating a user. Made a form for each one, it's been tested. Don't know how to fill attributes of CompanyProfile from the creation form, as I'm calling Company. Logic is kinda crazy, linked a video on discord of the dude explaining it. Maybe we can move every aspect of User to members/models.py, just to make it look better and have everything in its place. Left admin as only user.